Godfile data-store step 3: extract sqlite-workflow-repository
Move the workflow persistence cluster out of packages/data-store/src/sqlite-adapter.ts
into packages/data-store/src/sqlite-workflow-repository.ts. Workflow CRUD, snapshot and
rollup reads, findReviewGateByPr, and searchWorkflowsAndTasks now run over SqliteExecutor
in the new module; the adapter keeps one-line delegates. Behavior unchanged.
